+++ client/client-multi/client-c/src/kaa/kaa_user.c
@@ -476,7 +476,7 @@ kaa_error_t kaa_user_request_serialize(kaa_user_manager_t *self, kaa_platform_me
 
 kaa_error_t kaa_user_handle_server_sync(kaa_user_manager_t *self
                                       , kaa_platform_message_reader_t *reader
-                                      , uint32_t extension_options
+                                      , uint16_t extension_options
                                       , size_t extension_length)
 {
     KAA_RETURN_IF_NIL2(self, reader, KAA_ERR_BADPARAM);
